chore: clean up daft spark with better error handling
