+++ src/daft-connect/src/err.rs
@@ -7,9 +7,11 @@ macro_rules! invalid_argument_err {
 }
 
 #[macro_export]
-macro_rules! unimplemented_err {
-    ($arg: tt) => {{
-        let msg = format!($arg);
+macro_rules! nyi {
+    ($($arg:tt)*)  => {{
+        let msg = format!($($arg)*);
+        let msg = format!(r#"Feature: {msg} is not yet implemented, please open an issue at https://github.com/Eventual-Inc/Daft/issues/new?assignees=&labels=enhancement%2Cneeds+triage&projects=&template=feature_request.yaml"#);
+
         Err(::tonic::Status::unimplemented(msg))
     }};
 }
